# APP/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/groups/{groupId}/files/{fileId}": {
      "get": {
        "operationId": "FilesController_getFileMetadata",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadataResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ]
      }
    },
    "/api/groups/{groupId}/files/{fileId}/thumbnail": {
      "get": {
        "operationId": "FilesController_getFileThumbnail",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/gif": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ]
      }
    },
    "/api/groups/{groupId}/files/{fileId}/download-url": {
      "get": {
        "operationId": "FilesController_getFileDownloadUrl",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDownloadUrlResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ]
      }
    },
    "/api/groups/{groupId}/files/{fileId}/view-url": {
      "get": {
        "operationId": "FilesController_getFileViewUrl",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileViewUrlResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ]
      }
    },
    "/api/groups/{groupId}/files/{fileId}/content": {
      "get": {
        "operationId": "FilesController_getFileContent",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ]
      }
    },
...
    "/api/public/file-shares/unlock": {
      "post": {
        "operationId": "PublicFileSharesController_unlock",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockFileShareDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PublicFileShares"
        ]
      }
    },
    "/api/public/file-shares/{shareId}/content": {
      "head": {
        "operationId": "PublicFileSharesController_head",
        "parameters": [
          {
            "name": "shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicFileShares"
        ]
      },
      "get": {
        "operationId": "PublicFileSharesController_content",
        "parameters": [
          {
            "name": "shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicFileShares"
        ]
      }
    },
...
  },
  "info": {
    "title": "SekerChat HTTP API",
    "description": "Published contract for browser, CLI, reminder, and mobile clients.",
    "version": "0.8.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "seker_access"
      }
    },
    "schemas": {
...
      "FileMetadataResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "contentUrl": {
            "type": "string"
          },
          "metadataUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "uploaderId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "file"
            ]
          }
        },
        "required": [
          "id",
          "groupId",
          "originalName",
          "mimeType",
          "size",
          "createdAt",
          "contentUrl",
          "metadataUrl",
          "thumbnailUrl",
          "uploaderId",
          "kind"
        ]
      },
      "FileDownloadUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          }
        },
        "required": [
          "url",
          "originalName",
          "mimeType",
          "size"
        ]
      },
      "FileViewUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "mimeType",
          "size",
          "expiresAt"
        ]
      },
...
      "UnlockFileShareDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}